rag-service/requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
sentence-transformers==2.7.0
pinecone>=8.0.0
pydantic==2.8.2
requests==2.32.3
python-dotenv==1.0.1
groq>=0.10.0
prometheus-fastapi-instrumentator==5.11.1
llama-index==0.11.20
llama-index-core==0.11.20
